refactor: reorganize helpers folder
