+++ data_pipelines/elasticsearch/clean_data.py
