+++ data_pipelines/etl/data_fetch_clean/unite_legale.py
@@ -3,7 +3,7 @@ import shutil
 import logging
 import pandas as pd
 import requests
-from dag_datalake_sirene.utils.minio_helpers import (
+from dag_datalake_sirene.helpers.minio_helpers import (
     get_object_minio,
 )
 from dag_datalake_sirene.config import (
