+++ data_pipelines/etl/task_functions/create_dirigeants_tables.py
@@ -5,7 +5,7 @@ from dag_datalake_sirene.data_pipelines.etl.data_fetch_clean.dirigeants import (
     preprocess_dirigeants_pp,
 )
 
-from dag_datalake_sirene.utils.sqlite_client import SqliteClient
+from dag_datalake_sirene.helpers.sqlite_client import SqliteClient
 
 from dag_datalake_sirene.data_pipelines.etl.sqlite.helpers import (
     drop_table,
