+++ helpers/execute_slow_queries.py
@@ -2,7 +2,7 @@ import logging
 import requests
 
 from dag_datalake_sirene.config import AIO_URL
-from dag_datalake_sirene.utils.slow_requests import SLOW_REQUESTS
+from dag_datalake_sirene.helpers.slow_requests import SLOW_REQUESTS
 
 
 def execute_slow_requests():
