+++ maintenance/DAG_change_api_color.py
@@ -5,8 +5,8 @@ from airflow.models import DAG
 from airflow.operators.email_operator import EmailOperator
 from airflow.operators.python import PythonOperator
 from airflow.utils.dates import days_ago
-from dag_datalake_sirene.utils.get_colors import get_colors
-from dag_datalake_sirene.utils.update_color_file import update_color_file
+from dag_datalake_sirene.helpers.get_colors import get_colors
+from dag_datalake_sirene.helpers.update_color_file import update_color_file
 from dag_datalake_sirene.config import (
     AIRFLOW_ENV,
     EMAIL_LIST,
