+++ maintenance/DAG_flush_and_execute_queries.py
@@ -4,8 +4,8 @@ from airflow.models import DAG
 from airflow.operators.email_operator import EmailOperator
 from airflow.operators.python import PythonOperator
 from airflow.utils.dates import days_ago
-from dag_datalake_sirene.utils.flush_cache import flush_cache
-from dag_datalake_sirene.utils.execute_slow_elastic_queries import (
+from dag_datalake_sirene.helpers.flush_cache import flush_cache
+from dag_datalake_sirene.helpers.execute_slow_queries import (
     execute_slow_requests,
 )
 from dag_datalake_sirene.config import (
