+++ maintenance/DAG_flush_cache_only.py
@@ -4,7 +4,7 @@ from airflow.models import DAG
 from airflow.operators.email_operator import EmailOperator
 from airflow.operators.python import PythonOperator
 from airflow.utils.dates import days_ago
-from dag_datalake_sirene.utils.flush_cache import flush_cache
+from dag_datalake_sirene.helpers.flush_cache import flush_cache
 from dag_datalake_sirene.config import (
     AIRFLOW_ENV,
     EMAIL_LIST,
